Grant device access for console users instead of plugdev members.
Instead of having to add people to the plugdev group to talk to the device,
allow every locally logged in user acces using the uaccess udev tag.

It might be an alternative to keep the plugdev setup for backward compatibility
and to be able to grant access to background processes without a locally logged
in user.